integrating Yankuic's tests and migrations
