+++ survey/test_data/shp/complete/test_grid.shx
