+++ survey/test_data/shp/missing_prj/test_grid.shp
