+++ survey/test_data/shp/zipped/test_grid.zip
